(wip):small fixes
