+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
     
     strategy:
       matrix:
-        node-version: [ 10, 13 ]
+        node-version: [ 12 ]
     
     steps:
     - uses: actions/checkout@v2
