+++ package.json
@@ -35,6 +35,7 @@
     "sample": "ng serve accordion-sample -o --host 0.0.0.0 --port 8881 --live-reload",
     "sample:prod": "ng serve accordion-sample -o --host 0.0.0.0 --port 8882 --live-reload --prod",
     "clean:dist": "rimraf ./dist/**",
+    "clean:docs": "rimraf ./docs/**",
     "build": "ng b",
     "build:prod": "ng b --prod",
     "build:sample": "ng b accordion-sample --prod",
@@ -42,7 +43,9 @@
     "pack": "cd ./dist/@ngx-dummy/accordion-simple && npm pack",
     "publish": "cd ./dist/@ngx-dummy/accordion-simple && npm publish --access public",
     "publish:full": "run-s clean:dist build:prod pack publish",
-    "push:origin": "sh ./scripts/hooks/push.sh",
+    "gh-pages": "run-s build:sample -- --base-href https://vovansuper.github.io/accordion-simple/ cp:sample",
+    "prepush:origin": "run-p clean:docs gh-pages",
+    "push:origin": "sh ./scripts/push.sh",
     "test": "ng test",
     "test:ci": "ng test --karma-config=karma-ci.conf.js --no-watch --no-progress --browsers=ChromeHeadlessCI"
   },
@@ -70,7 +73,6 @@
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "^13.11.1",
     "chalk": "^4.0.0",
-    "child-process-promise": "^2.2.1",
     "codelyzer": "^5.2.2",
     "husky": "^4.2.5",
     "jasmine-core": "^3.5.0",
@@ -80,13 +82,14 @@
     "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "^3.1.1",
     "karma-jasmine-html-reporter": "^1.5.3",
-    "ng-packagr": "^9.1.0",
+    "ng-packagr": "^9.1.1",
+    "node-jq": "^1.11.0",
     "npm-run-all": "^4.1.5",
     "ora": "^4.0.3",
     "protractor": "^5.4.3",
     "rimraf": "^3.0.2",
     "simple-git": "^1.132.0",
-    "svg-to-ts": "^4.1.1",
+    "svg-to-ts": "^4.2.1",
     "ts-node": "^8.8.2",
     "tslint": "^6.1.1",
     "typescript": "~3.8.3"
