+++ scripts/versioning/bump.js
@@ -1,3 +1,5 @@
+
+const d = console.debug;
 function udpateMajorVersion(version) {
 	return updateVersion(version, 'major');
 }
@@ -7,6 +9,28 @@ function updateMinorVersion(version) {
 function updatePatchVersion(version) {
 	return updateVersion(version, 'patch');
 }
+
+
+const verReg = /(["|']version["|']\s?\:\s*)\"([\.\d]+)"/gi;
+/**
+ * 
+ * @param {string} packageJson - contents of 'package.json' file to be parsed / version field updated
+ * @param {string} newVersion - string representations of new 'version' field to be injected in 'package.json' file (i.e, "1.1.0")
+ */
+const updateVersion = (packageJson, newVersion) => {
+
+  // const veres2018Reg = /(?<="version":\s?)(\.\d)+/;
+
+  let version = packageJson.match(verReg);
+  d('PackageVersion capture: ', version);
+
+  const newVer = JSON.stringify(newVersion);
+  packageJson = packageJson.replace(verReg, `$1${newVer}`);
+
+  d('Version after: ', packageJson);
+
+  return packageJson;
+};
 
 /**
  *
@@ -33,5 +57,6 @@ function updateVersion(version, type) {
 module.exports = {
 	udpateMajorVersion,
 	updateMinorVersion,
-	updatePatchVersion
+	updatePatchVersion,
+	updateVersion
 };
